Modify MockService

+++ src/main/java/com/ajou/TwotoSix/unitTest2/service/MockService.java
@@ -19,13 +19,13 @@ public class MockService {
         return student;
     }
 
-    public double updateGPAByStudentId(String studentId, float updatedGPA) {
+    public double updateGPAByStudentId(String studentId, double updatedGPA) {
         Student student = findBystudentId(studentId);
         student.setGPA(updatedGPA);
         return student.getGPA();
     }
   
-    public Student addStudent(String name, String studentId, int currentSemester, String major, float GPA){
+    public Student addStudent(String name, String studentId, int currentSemester, String major, double GPA){
         Student addStudent = new Student(name, studentId, currentSemester, major, GPA);
         return addStudent;
     }
